edit api to work with db
change kIndId from byte to int

use ki insted of ki1 in sim card table

+++ CommunicationCoverageSupport.BLL/Services/ISimCardDrainService.cs
@@ -13,6 +13,6 @@ namespace CommunicationCoverageSupport.BLL.Services
         Task<IEnumerable<SimCardDrainDto>> GetAllAsync();
         Task<SimCardDrainDto?> GetByIccidAsync(string iccid);
         Task<SimCardDrainFullInfoDto?> GetFullInfoByIccidAsync(string iccid);
-        Task<bool> DeleteAsync(string iccid, string imsi, string msisdn, byte kIndId);
+        Task<bool> DeleteAsync(string iccid, string imsi, string msisdn, int kIndId);
     }
 }
