+++ CommunicationCoverageSupport.BLL/Services/ISimCardService.cs
@@ -10,6 +10,6 @@ namespace CommunicationCoverageSupport.BLL.Services.SimCards
         Task<SimCardFullInfoDto?> GetFullInfoByIccidAsync(string iccid);
         Task<bool> CreateAsync(SimCardDto dto);
         Task<bool> UpdateAsync(SimCardDto dto);
-        Task<string> DrainAsync(string iccid, string imsi, string msisdn, byte kIndId);
+        Task<string> DrainAsync(string iccid, string imsi, string msisdn, int kIndId);
     }
 }
