+++ CommunicationCoverageSupport.BLL/Services/SimCardDrainService.cs
@@ -21,6 +21,6 @@ namespace CommunicationCoverageSupport.BLL.Services
         public Task<IEnumerable<SimCardDrainDto>> GetAllAsync() => _repository.GetAllAsync();
         public Task<SimCardDrainDto?> GetByIccidAsync(string iccid) => _repository.GetByIccidAsync(iccid);
         public Task<SimCardDrainFullInfoDto?> GetFullInfoByIccidAsync(string iccid) => _repository.GetFullInfoByIccidAsync(iccid);
-        public Task<bool> DeleteAsync(string iccid, string imsi, string msisdn, byte kIndId) => _repository.DeleteAsync(iccid, imsi, msisdn, kIndId);
+        public Task<bool> DeleteAsync(string iccid, string imsi, string msisdn, int kIndId) => _repository.DeleteAsync(iccid, imsi, msisdn, kIndId);
     }
 }
